feat: profile points system (take II).

+++ apps/services/pkg/api/adapters/storage/repository_profiles.go
@@ -151,7 +151,7 @@ func (r *Repository) GetProfileByID(
 		CreatedAt:         row.Profile.CreatedAt,
 		UpdatedAt:         vars.ToTimePtr(row.Profile.UpdatedAt),
 		DeletedAt:         vars.ToTimePtr(row.Profile.DeletedAt),
-		Points:            int(row.Profile.Points),
+		Points:            uint64(row.Profile.Points),
 	}
 
 	return result, nil
@@ -380,7 +380,7 @@ func (r *Repository) ListProfileContributions( //nolint:funlen
 				CreatedAt:         row.Profile_2.CreatedAt,
 				UpdatedAt:         vars.ToTimePtr(row.Profile_2.UpdatedAt),
 				DeletedAt:         vars.ToTimePtr(row.Profile_2.DeletedAt),
-				Points:            int(row.Profile_2.Points),
+				Points:            uint64(row.Profile_2.Points),
 			},
 		}
 	}
@@ -453,7 +453,7 @@ func (r *Repository) ListProfileMembers(
 				CreatedAt:         row.Profile_2.CreatedAt,
 				UpdatedAt:         vars.ToTimePtr(row.Profile_2.UpdatedAt),
 				DeletedAt:         vars.ToTimePtr(row.Profile_2.DeletedAt),
-				Points:            int(row.Profile_2.Points),
+				Points:            uint64(row.Profile_2.Points),
 			},
 		}
 	}
