+++ apps/services/pkg/api/business/profiles/types.go
@@ -27,7 +27,7 @@ type Profile struct {
 	Kind              string     `json:"kind"`
 	Title             string     `json:"title"`
 	Description       string     `json:"description"`
-	Points            int        `json:"points"`
+	Points            uint64     `json:"points"`
 }
 
 type ProfileWithChildren struct {
